requirements.txt
asgiref==3.10.0
Django==5.2.7
django-modeltranslation==0.18.11
pillow==11.3.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
typing_extensions==4.15.0
